Move DataExtractor interface to ML plugin
